+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `URLMap` is exported from `veloce` and is the public name of the class `Veloce.url_map` returns. ([#288](https://github.com/Lokesh-Tallapaneni/veloce/pull/288))
 - `WebSocketState` is exported from `veloce`; it is the declared return type of `WebSocket.application_state` and `.client_state`. ([#288](https://github.com/Lokesh-Tallapaneni/veloce/pull/288))
 
 ### Changed